wip: not replaying previous RMS
